requirements.txt
aiobotocore 
aiohappyeyeballs 
aiohttp 
aioitertools 
aiosignal 
alabaster 
altair 
anaconda-anon-usage 
anaconda-auth 
anaconda-catalogs 
anaconda-cli-base 
anaconda-client 
anaconda-navigator 
anaconda-project 
annotated-types 
anyio 
appdirs==1.4.4
archspec 
argon2-cffi 
argon2-cffi-bindings 
arrow 
astroid 
astropy 
astropy-iers-data 
asttokens 
async-lru 
asyncssh 
atomicwrites==1.4.0
attrs 
Automat 
autopep8 
babel 
bcrypt 
beautifulsoup4 
binaryornot 
black 
bleach 
blinker 
bokeh 
boltons 
botocore 
Bottleneck 
Brotli 
cachetools 
certifi 
cffi 
chardet 
charset-normalizer 
click 
cloudpickle 
colorama 
colorcet 
comm 
conda 
conda-anaconda-telemetry 
conda-anaconda-tos 
conda-build 
conda-content-trust 
conda-libmamba-solver 
conda-pack 
conda-package-handling 
conda-repo-cli 
conda-token 
conda_index 
conda_package_streaming 
constantly 
contourpy 
cookiecutter 
cryptography 
cssselect 
cycler 
cytoolz 
dask 
dask-expr 
datashader 
debugpy 
decorator 
defusedxml 
Deprecated 
diff-match-patch 
dill 
distributed 
distro 
docstring-to-markdown 
docutils 
et_xmlfile 
evalidate 
executing 
fastjsonschema 
filelock 
flake8 
Flask 
fonttools 
frozendict 
frozenlist 
fsspec 
gitdb 
GitPython 
gmpy2 
greenlet 
h11 
h5py 
HeapDict 
holoviews 
httpcore 
httpx 
hvplot 
hyperlink 
idna 
imageio 
imagesize 
imbalanced-learn 
importlib_metadata 
incremental 
inflection 
iniconfig 
intake 
intervaltree 
ipykernel 
ipython 
ipywidgets 
isort 
itemadapter 
itemloaders 
itsdangerous 
jaraco.classes 
jaraco.context 
jaraco.functools 
jedi 
jeepney 
jellyfish 
Jinja2 
jinja2-time 
jmespath 
joblib 
json5 
jsonpatch 
jsonpointer==2.1
jsonschema 
jsonschema-specifications 
jupyter 
jupyter-console 
jupyter-events 
jupyter-lsp 
jupyter_client 
jupyter_core 
jupyter_server 
jupyter_server_terminals 
jupyterlab 
jupyterlab_pygments 
jupyterlab_server 
jupyterlab_widgets 
keyring 
kiwisolver 
lazy_loader 
lckr_jupyterlab_variableinspector 
libarchive-c 
libmambapy 
lief 
linkify-it-py 
llvmlite 
lmdb 
locket 
lxml 
lz4 
Markdown 
markdown-it-py 
MarkupSafe 
matplotlib==3.10.0
matplotlib-inline 
mccabe 
mdit-py-plugins 
mdurl 
menuinst 
mistune 
mkl-service==2.4.0
mkl_fft 
mkl_random 
more-itertools 
mpmath 
msgpack 
multidict 
multipledispatch 
mypy 
mypy_extensions 
narwhals 
navigator-updater 
nbclient 
nbconvert 
nbformat 
nest_asyncio 
networkx 
nltk 
notebook 
notebook_shim 
numba 
numexpr 
numpy 
numpydoc 
openpyxl 
overrides 
packaging 
pandas 
pandocfilters 
panel 
param 
parsel 
parso 
partd 
pathspec 
patsy 
pexpect 
pickleshare 
pillow 
pkce 
pkginfo 
platformdirs 
plotly 
pluggy 
ply 
poyo 
prometheus_client 
prompt_toolkit 
propcache 
Protego 
protobuf 
psutil 
ptyprocess 
pure-eval 
py-cpuinfo 
pyarrow 
pyasn1 
pyasn1-modules==0.2.8
pycodestyle 
pycosat 
pycparser 
pyct 
pycurl 
pydantic 
pydantic-settings 
pydantic_core 
PyDispatcher 
pydocstyle 
pyerfa 
pyflakes 
PyGithub 
Pygments 
PyJWT 
pylint 
pylint-venv 
pyls-spyder==0.4.0
PyNaCl 
pyodbc 
pyOpenSSL 
pyparsing 
PyQt5==5.15.10
PyQt5_sip 
PyQtWebEngine==5.15.6
PySocks 
pytest 
python-dateutil 
python-dotenv 
python-json-logger 
python-lsp-black 
python-lsp-jsonrpc 
python-lsp-server 
python-slugify 
pytoolconfig 
pytz 
pyuca 
pyviz_comms 
PyWavelets 
pyxdg 
PyYAML 
pyzmq 
QDarkStyle 
qstylizer 
QtAwesome 
qtconsole 
QtPy 
queuelib 
readchar 
referencing 
regex 
requests 
requests-file 
requests-toolbelt 
rfc3339_validator 
rfc3986_validator 
rich 
roman-numerals-py 
rope 
rpds-py 
Rtree 
ruamel.yaml 
ruamel.yaml.clib 
ruamel_yaml_conda 
s3fs 
scikit-image 
scikit-learn 
scipy 
Scrapy 
seaborn 
SecretStorage 
semver 
Send2Trash 
service-identity 
setuptools==72.1.0
shellingham 
sip 
six 
sklearn-compat 
smmap 
sniffio 
snowballstemmer 
sortedcontainers 
soupsieve 
Sphinx 
sphinxcontrib-applehelp 
sphinxcontrib-devhelp 
sphinxcontrib-htmlhelp 
sphinxcontrib-jsmath 
sphinxcontrib-qthelp 
sphinxcontrib-serializinghtml 
spyder 
spyder-kernels 
SQLAlchemy 
stack-data 
statsmodels 
streamlit 
superqt 
sympy 
tables 
tabulate 
tblib 
tenacity 
terminado 
text-unidecode 
textdistance 
threadpoolctl 
three-merge 
tifffile 
tinycss2 
tldextract 
toml 
tomli 
tomlkit 
toolz 
tornado 
tqdm 
traitlets 
truststore 
Twisted 
typer 
typing-inspection 
typing_extensions 
tzdata 
uc-micro-py 
ujson 
Unidecode 
urllib3 
w3lib 
watchdog 
wcwidth 
webencodings 
websocket-client 
Werkzeug 
whatthepatch 
wheel==0.45.1
widgetsnbextension 
wrapt 
wurlitzer 
xarray 
xyzservices 
yapf 
yarl 
zict 
zipp 
zope.interface 
zstandard 
